final rough draft of control plane changes

+++ src/control/deleteCollection.ts
@@ -1,4 +1,5 @@
 import { ManageIndexesApi } from '../pinecone-generated-ts-fetch/control';
+import { PineconeArgumentError } from '../errors';
 // import { buildConfigValidator } from '../validator';
 // import { CollectionNameSchema } from './types';
 // import type { CollectionName } from './types';
@@ -6,16 +7,25 @@ import { ManageIndexesApi } from '../pinecone-generated-ts-fetch/control';
 /**
  * The name of collection to delete.
  */
-export type DeleteCollectionOptions = CollectionName;
+// export type DeleteCollectionOptions = CollectionName;
 
 export const deleteCollection = (api: ManageIndexesApi) => {
-  const validator = buildConfigValidator(
-    CollectionNameSchema,
-    'deleteCollection'
-  );
+  // const validator = buildConfigValidator(
+  //   CollectionNameSchema,
+  //   'deleteCollection'
+  // );
 
-  return async (collectionName: DeleteCollectionOptions): Promise<void> => {
-    validator(collectionName);
+  return async (collectionName: string): Promise<void> => {
+    if (
+      !collectionName ||
+      collectionName.length === 0 ||
+      typeof collectionName !== 'string'
+    ) {
+      throw new PineconeArgumentError(
+        'You must enter a non-empty string for the `collectionName` field.'
+      );
+    }
+    // validator(collectionName);
 
     await api.deleteCollection({ collectionName });
     return;
