+++ src/control/deleteIndex.ts
@@ -1,15 +1,13 @@
 import { ManageIndexesApi } from '../pinecone-generated-ts-fetch/control';
-import { buildConfigValidator } from '../validator';
-import { IndexName, IndexNameSchema } from './types';
-
-/** The name of index to delete */
-export type DeleteIndexOptions = IndexName;
+import { PineconeArgumentError } from '../errors';
 
 export const deleteIndex = (api: ManageIndexesApi) => {
-  const validator = buildConfigValidator(IndexNameSchema, 'deleteIndex');
-
-  return async (indexName: DeleteIndexOptions): Promise<void> => {
-    validator(indexName);
+  return async (indexName: string): Promise<void> => {
+    if (!indexName || indexName.length === 0 || typeof indexName !== 'string') {
+      throw new PineconeArgumentError(
+        'You must enter a non-empty string for the `indexName` field.'
+      );
+    }
 
     await api.deleteIndex({ indexName });
     return;
